backend/requirements.txt
Django==1.11.5
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
Markdown==2.6.9
PyJWT==1.5.3
pytz==2017.2
